updating the recognition program
add CLAHE to the program
